Move gui.cpp and gui.h to software root.
